+++ tests/test_table/test_fuse.py
@@ -1,0 +1,81 @@
+#!/usr/bin/env python
+# -*- coding: utf8 -*-
+
+from agate import Table
+from agate.data_types import *
+from agate.testcase import AgateTestCase
+
+
+class TestFuse(AgateTestCase):
+    def setUp(self):
+        self.left_rows = (
+            (1, 4, 'a'),
+            (2, 3, 'b'),
+            (None, 2, 'c')
+        )
+
+        self.right_rows = (
+            (1, 4, 'a'),
+            (2, 3, 'b'),
+            (None, 2, 'c')
+        )
+
+        self.number_type = Number()
+        self.text_type = Text()
+
+        self.left_column_names = ['one', 'two', 'three']
+        self.right_column_names = ['four', 'five', 'six']
+        self.column_types = [self.number_type, self.number_type, self.text_type]
+
+        self.left = Table(self.left_rows, self.left_column_names, self.column_types)
+        self.right = Table(self.right_rows, self.right_column_names, self.column_types)
+
+    def test_fuse(self):
+        new_table = self.left.fuse(self.right)
+
+        self.assertIsNot(new_table, self.left)
+        self.assertIsNot(new_table, self.right)
+        self.assertColumnNames(new_table, ['one', 'two', 'three', 'four', 'five', 'six'])
+        self.assertColumnTypes(new_table, [Number, Number, Text, Number, Number, Text])
+        self.assertRows(new_table, [
+            (1, 4, 'a', 1, 4, 'a'),
+            (2, 3, 'b', 2, 3, 'b'),
+            (None, 2, 'c', None, 2, 'c')
+        ])
+
+    def test_fuse_short_right(self):
+        right_rows = self.right_rows + ((7, 9, 'z'),)
+        right = Table(right_rows, self.right_column_names, self.column_types)
+
+        new_table = self.left.fuse(right)
+
+        self.assertIsNot(new_table, self.left)
+        self.assertIsNot(new_table, right)
+        self.assertColumnNames(new_table, ['one', 'two', 'three', 'four', 'five', 'six'])
+        self.assertColumnTypes(new_table, [Number, Number, Text, Number, Number, Text])
+        self.assertRows(new_table, [
+            (1, 4, 'a', 1, 4, 'a'),
+            (2, 3, 'b', 2, 3, 'b'),
+            (None, 2, 'c', None, 2, 'c'),
+            (None, None, None, 7, 9, 'z')
+        ])
+
+    def test_fuse_short_left(self):
+        left_rows = self.left_rows + ((7, 9, 'z'),)
+        left = Table(left_rows, self.left_column_names, self.column_types)
+
+        new_table = left.fuse(self.right)
+
+        import sys
+        new_table.print_table(output=sys.stdout)
+
+        self.assertIsNot(new_table, left)
+        self.assertIsNot(new_table, self.right)
+        self.assertColumnNames(new_table, ['one', 'two', 'three', 'four', 'five', 'six'])
+        self.assertColumnTypes(new_table, [Number, Number, Text, Number, Number, Text])
+        self.assertRows(new_table, [
+            (1, 4, 'a', 1, 4, 'a'),
+            (2, 3, 'b', 2, 3, 'b'),
+            (None, 2, 'c', None, 2, 'c'),
+            (7, 9, 'z', None, None, None)
+        ])